% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurvData.R
\name{recordstatus}
\alias{recordstatus}
\title{to creat a record of censoring event into numeric value}
\usage{
recordstatus(x)
}
\arguments{
\item{x}{event}
}
\value{
return the result if the x event is censored
}
\description{
A function to convert the censoring event into numeric
value for surivival analysis
}
